+++ RELEASES.md
@@ -878,6 +878,7 @@
 -----------------
   - fix: SMC-3242 - Will correctly listRelevantPermissions for onBehalfOf sessions
 
+
 11.13.3 2021-09-08
 -----------------
   - fix: SMC-4044 - cookie events not dependent on client login
@@ -895,6 +896,17 @@
 -----------------
   - fix: SMC-4512 -  data provider get action causes fatal when allowNestedPermissions switched on
 
-11.X.X 2021-11-16
+11.14.0 2021-11-24
+-----------------
+  - feature: SMC-2954 - Allows for configuration and use of multiple authentication providers.
+
+11.14.1 2021-11-24
+-----------------
+  - fix: SMC-4386 -  upsertMultiplePermissions will now allow for removing permissions/prohibitions as well as upserting permissions or prohibitions
+
+11.14.2 2021-11-26
 -----------------
   - fix: SMC-2716 -  fatal caused when message passed to processAuthorize (security service) has no message.request.path
+  - fixed test was leaking clients
+
+